rename files input_parameter.h
