+++ include/exadg/compressible_navier_stokes/spatial_discretization/kernels_and_operators.h
@@ -30,7 +30,7 @@
 
 // ExaDG
 #include <exadg/compressible_navier_stokes/user_interface/boundary_descriptor.h>
-#include <exadg/compressible_navier_stokes/user_interface/input_parameters.h>
+#include <exadg/compressible_navier_stokes/user_interface/parameters.h>
 #include <exadg/functions_and_boundary_conditions/evaluate_functions.h>
 #include <exadg/matrix_free/integrators.h>
 #include <exadg/operators/interior_penalty_parameter.h>
