+++ include/exadg/compressible_navier_stokes/user_interface/parameters.cpp
@@ -23,7 +23,7 @@
 #include <deal.II/base/exceptions.h>
 
 // ExaDG
-#include <exadg/compressible_navier_stokes/user_interface/input_parameters.h>
+#include <exadg/compressible_navier_stokes/user_interface/parameters.h>
 
 namespace ExaDG
 {
