+++ include/exadg/convection_diffusion/spatial_discretization/operators/convective_operator.h
@@ -23,7 +23,7 @@
 #define CONV_DIFF_CONVECTION_OPERATOR
 
 #include <exadg/convection_diffusion/user_interface/boundary_descriptor.h>
-#include <exadg/convection_diffusion/user_interface/input_parameters.h>
+#include <exadg/convection_diffusion/user_interface/parameters.h>
 #include <exadg/functions_and_boundary_conditions/evaluate_functions.h>
 #include <exadg/matrix_free/integrators.h>
 #include <exadg/operators/operator_base.h>
