+++ include/exadg/convection_diffusion/spatial_discretization/operators/diffusive_operator.h
@@ -23,7 +23,7 @@
 #define CONV_DIFF_DIFFUSIVE_OPERATOR
 
 #include <exadg/convection_diffusion/user_interface/boundary_descriptor.h>
-#include <exadg/convection_diffusion/user_interface/input_parameters.h>
+#include <exadg/convection_diffusion/user_interface/parameters.h>
 #include <exadg/operators/interior_penalty_parameter.h>
 #include <exadg/operators/operator_base.h>
