+++ include/exadg/convection_diffusion/time_integration/create_time_integrator.h
@@ -24,7 +24,7 @@
 
 #include <exadg/convection_diffusion/time_integration/time_int_bdf.h>
 #include <exadg/convection_diffusion/time_integration/time_int_explicit_runge_kutta.h>
-#include <exadg/convection_diffusion/user_interface/input_parameters.h>
+#include <exadg/convection_diffusion/user_interface/parameters.h>
 
 namespace ExaDG
 {
