+++ include/exadg/incompressible_flow_with_transport/user_interface/application_base.h
@@ -26,7 +26,7 @@
 #include <exadg/convection_diffusion/postprocessor/postprocessor.h>
 #include <exadg/convection_diffusion/user_interface/boundary_descriptor.h>
 #include <exadg/convection_diffusion/user_interface/field_functions.h>
-#include <exadg/convection_diffusion/user_interface/input_parameters.h>
+#include <exadg/convection_diffusion/user_interface/parameters.h>
 #include <exadg/grid/grid.h>
 #include <exadg/incompressible_navier_stokes/user_interface/application_base.h>
