+++ include/exadg/incompressible_navier_stokes/spatial_discretization/spatial_operator_base.h
@@ -46,7 +46,7 @@
 #include <exadg/incompressible_navier_stokes/spatial_discretization/turbulence_model.h>
 #include <exadg/incompressible_navier_stokes/user_interface/boundary_descriptor.h>
 #include <exadg/incompressible_navier_stokes/user_interface/field_functions.h>
-#include <exadg/incompressible_navier_stokes/user_interface/input_parameters.h>
+#include <exadg/incompressible_navier_stokes/user_interface/parameters.h>
 #include <exadg/matrix_free/matrix_free_data.h>
 #include <exadg/operators/inverse_mass_operator.h>
 #include <exadg/operators/mass_operator.h>
