+++ include/exadg/incompressible_navier_stokes/spatial_discretization/turbulence_model.h
@@ -27,7 +27,7 @@
 
 // ExaDG
 #include <exadg/incompressible_navier_stokes/spatial_discretization/operators/viscous_operator.h>
-#include <exadg/incompressible_navier_stokes/user_interface/input_parameters.h>
+#include <exadg/incompressible_navier_stokes/user_interface/parameters.h>
 #include <exadg/matrix_free/integrators.h>
 
 namespace ExaDG
