+++ include/exadg/incompressible_navier_stokes/user_interface/parameters.cpp
@@ -23,7 +23,7 @@
 #include <deal.II/base/exceptions.h>
 
 // ExaDG
-#include <exadg/incompressible_navier_stokes/user_interface/input_parameters.h>
+#include <exadg/incompressible_navier_stokes/user_interface/parameters.h>
 #include <exadg/utilities/print_functions.h>
 
 namespace ExaDG
