+++ include/exadg/structure/spatial_discretization/operator.h
@@ -38,7 +38,7 @@
 #include <exadg/structure/spatial_discretization/operators/nonlinear_operator.h>
 #include <exadg/structure/user_interface/boundary_descriptor.h>
 #include <exadg/structure/user_interface/field_functions.h>
-#include <exadg/structure/user_interface/input_parameters.h>
+#include <exadg/structure/user_interface/parameters.h>
 
 namespace ExaDG
 {
